Rename BulkPrice to SingleAndBulkUnitPrice
